Fix: Handle undefined window by returning silently
in some apps (like nextjs apps) you may have tests that run in node jest environment and others that run in jsdom. This can be configured with a comment on each file. When this hook is applied globally, it blows up if a particular test file switches to the node environment. But with this change we can silently just dont do anything in node environments.